[#11] Refactor: ServiceImpl 분리 제거
