+++ src/test/java/sparta/com/sappun/domain/user/controller/UserControllerTest.java
@@ -1,6 +1,5 @@
 package sparta.com.sappun.domain.user.controller;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
