+++ src/test/java/sparta/com/sappun/domain/user/service/impl/UserServiceTest.java
@@ -22,16 +22,17 @@ import sparta.com.sappun.domain.user.dto.response.UserLoginRes;
 import sparta.com.sappun.domain.user.entity.Role;
 import sparta.com.sappun.domain.user.entity.User;
 import sparta.com.sappun.domain.user.repository.UserRepository;
+import sparta.com.sappun.domain.user.service.UserService;
 import sparta.com.sappun.global.exception.GlobalException;
 import sparta.com.sappun.test.UserTest;
 
 @ExtendWith(MockitoExtension.class)
-class UserServiceImplTest implements UserTest {
+class UserServiceTest implements UserTest {
     @Mock UserRepository userRepository;
 
     @Mock PasswordEncoder passwordEncoder;
 
-    @InjectMocks UserServiceImpl userService;
+    @InjectMocks UserService userService;
 
     @Captor ArgumentCaptor<User> argumentCaptor;
